Implemented initial support for font styling.
Notekata
- Added support to NKTTextView to set text attributes on selected text ranges
- Adding typing attributes to NKTTextView that help out in determining the style of text that is being edited
- Added NKTFontViewController, which is used to selected font family names
- Fixed a ton of UI bugs and issues

KobaCore
- Added more KBCClamp functions

KobaText
- Refactored KBTTextStyleGenerator into KBTStyleDescriptor and KBTFontFamilyDescriptor
- Added functions that query a given font name for traits that is supports
- Added more diagnostic functions

KobaUI
- Set disabled state colors for KUIToggleButton

+++ Classes/NKTTextSection.m
@@ -212,7 +212,7 @@
     firstLineIndex -= numberOfSkirtLines_;
     firstLineIndex = MAX(firstLineIndex, 0);
     
-    if (firstLineIndex > [typesettedLines_ count] - 1)
+    if (firstLineIndex > (NSInteger)([typesettedLines_ count] - 1))
     {
         return NSMakeRange(NSNotFound, 0);
     }
